[refactor] #99 뒤로가기 버튼 리팩토링
